requirements.txt
requests == 2.25.1
pytz == 2021.1
boto3 == 1.17.59
beautifulsoup4 == 4.9.3
xmltodict == 0.12.0
python-dotenv == 0.19.2